renamed aggregator.rs to router.rs
